pyproject.toml
[tool.poetry]
name = "wanteat-be"
version = "0.1.0"
description = "list of dependencies installed in wanteat"
authors = ["Matteo Meneguzzi <matteo.dev.meneguzzi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pymongo = "^4.6.1"
mongoengine = "^0.27.0"
tqdm = "^4.66.2"
colorama = "^0.4.6"
python-dateutil = "^2.8.2"
motor = "^3.3.2"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.1"
uuid = "^1.30"
pydantic = {extras = ["email"], version = "^2.6.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
